+++ src/components/BorrowerProfile/LoanStory.vue
@@ -77,6 +77,7 @@ export default {
 							hash
 						}
 						... on Video {
+							id
 							youtubeId
 						}
 					}