+++ src/components/GoalSetting/GoalSettingContainer.vue
@@ -82,7 +82,6 @@
 					:loading-current-year="loadingCurrentYear"
 					:goal-progress="goalProgress"
 					:is-goal-completed="isGoalCompleted"
-					show-goal-value-props-copy
 					use-direct-question-title
 					@set-goal-target="setTarget($event)"
 					@set-goal="setGoal($event)"