+++ src/components/MyKiva/ImpactInsight/MyKivaImpactInsightScreen2.vue
@@ -7,11 +7,11 @@
 			>
 				<div
 					class="tw-flex tw-items-center tw-justify-center tw-rounded-full tw-border-4 tw-border-marigold
-							tw-bg-caution-highlight tw-text-center min-big-circle-size"
+								tw-bg-caution-highlight tw-text-center min-big-circle-size"
 				>
 					<div
 						class="tw-flex tw-items-center tw-justify-center tw-rounded-full tw-border-4 tw-border-brand-500
-							tw-text-center min-circle-size"
+								tw-text-center min-circle-size"
 					>
 						<div class="tw-items-center tw-gap-3">
 							<div
@@ -22,7 +22,7 @@
 								/>
 							</div>
 
-							<div class="tw-text-lg tw-font-semibold tw-text-slate-900">
+							<div class="tw-text-base tw-text-slate-900">
 								<strong>Their loan</strong>
 							</div>
 						</div>
@@ -33,7 +33,7 @@
 				>
 					<p
 						class="tw-rounded-2xl tw-bg-slate-100 tw-py-2 tw-px-2 md:!tw-px-3
-							tw-text-base tw-leading-relaxed tw-bg-gray-100 tw-rounded-md md:tw-text-lg"
+								tw-text-base tw-bg-gray-100 tw-rounded-md"
 					>
 						<strong class="tw-text-brand">This loan is special because</strong>
 						{{ whySpecial }}